renaming av funksjoner og components
