+++ frontend/mulighetsrommet-veileder-flate/src/components/filtrering/FilterTagsContainer.tsx
@@ -1,0 +1,10 @@
+import styles from "@/components/filtrering/FilterTagsContainer.module.scss";
+import { PropsWithChildren } from "react";
+
+export function FilterTagsContainer(props: PropsWithChildren) {
+  return (
+    <div className={styles.filtertags} data-testid="filtertags">
+      {props.children}
+    </div>
+  );
+}
